+++ tests/test_open.py
@@ -238,6 +238,10 @@ def test_opening_model(tmp_path, node_class):
             assert model.asn_type == "image"
         elif node_class == stnode.WfiMosaic:
             assert model.meta.basic.optical_element == "F158"
+        elif node_class in (stnode.SegmentationMap, stnode.SourceCatalog):
+            assert model.meta.optical_element == "F158"
+        elif node_class in (stnode.MosaicSegmentationMap, stnode.MosaicSourceCatalog):
+            assert hasattr(model.meta, "basic")
         else:
             assert model.meta.instrument.optical_element == "F158"